pyproject.toml
[tool.poetry]
name = "jamerbot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.0.1"
flask = "^1.1.2"
praw = "^7.1.4"
python-dotenv = "^0.15.0"
redditcli = "^2.1.0"
wikipedia = "^1.4.0"
youtube-dl = "^2021.2.4"
aiohttp = "^3.7.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
